Rename colList, dark mode icon fix, added role to user table
Collections_latest renamed to CollectionsList
Dark/light mode icon was not showing the proper icon now its fixed
Added colored roles in the user table
